test: simplify topic reader metric coverage
